check if file exists to prevent win issue

+++ src/annoylib.h
@@ -833,7 +833,9 @@ public:
 
   bool save(const char* filename, bool prefault=false) {
     // Delete file if it already exists (See issue #335)
-    unlink(filename);
+    struct stat buffer;
+    if (stat(filename, &buffer) == 0)
+      unlink(filename);
 
     FILE *f = fopen(filename, "wb");
     if (f == NULL)
